liblights: Fix case where liblight is already defined
Display-caf-8916 recently got an update for HW color metaData etc.
This added the include of liblights to be targeted as a shared
library. The issue is, common repos have been calling liblights
for awhile now with their own modified implementation of liblights.

This causes: hardware/qcom/display-caf/msm8916/liblight:
             MODULE.TARGET.SHARED_LIBRARIES.lights.msm8916
             already defined by device/motorola/msm8916-common/liblight.

The only solution i can find right now is to not include it as a shared library
from msm8916-common. The include exist in Display-caf already. Probably not the
best solution but it doees fix compilation. Runtime still needs testing.

Change-Id: I27bf4749aa53612d401e7930ed52d21bc5189b69